Fix lensfun data dir
Darktable looks for the lensfun database in the parent directory of its
own data dir. The Ubuntu liblensfun-data-v1 deb installs this data under
`usr/share`, while the cmake plugin installs darktable's data to `share`
by default.

This moves the darktable data dir under `usr` so that it is able to find
the installed lensfun data.